[B2BQUOTES-42] Store app settings in vbase, reorganize code
