+++ node/clients/checkout.ts
@@ -5,11 +5,10 @@ import type {
 } from '@vtex/api'
 import { JanusClient } from '@vtex/api'
 
+import { APP_NAME } from '../constants'
 import { createTracing } from '../utils/index'
 
 const CHECKOUT_ENDPOINT = 'api/checkout/pvt/configuration/orderForm'
-
-export const CHECKOUT_APP = 'b2b-quotes-graphql'
 
 export default class Checkout extends JanusClient {
   constructor(ctx: IOContext, options?: InstanceOptions) {
@@ -26,7 +25,7 @@ export default class Checkout extends JanusClient {
   }
 
   public getOrderFormConfiguration(tracingConfig?: RequestTracingConfig) {
-    const metric = `${CHECKOUT_APP}-getOrderForm`
+    const metric = `${APP_NAME}-getOrderForm`
 
     return this.http.get<OrderFormConfiguration>(CHECKOUT_ENDPOINT, {
       metric,
@@ -39,7 +38,7 @@ export default class Checkout extends JanusClient {
     userToken: string,
     tracingConfig?: RequestTracingConfig
   ) {
-    const metric = `${CHECKOUT_APP}-setOrderForm`
+    const metric = `${APP_NAME}-setOrderForm`
 
     return this.http.post(CHECKOUT_ENDPOINT, body, {
       headers: {
